+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Post v1 a change is further considered breaking if the state file format is inco
 
 ### Upcoming
 
+### v1.3.0.0
+
 *Major*
 
 * Added activity field to Role. ([#231](https://github.com/hjwylde/werewolf/issues/231))